Create Cube3D class to handle permutations
